Refactor Select component stories
